worker: Move contents of WorkerForBuilderBase to deriving classes
